+++ dev/sparktestsupport/modules.py
@@ -743,10 +743,10 @@ pyspark_pandas = Module(
         "pyspark.pandas.tests.test_repr",
         "pyspark.pandas.tests.resample.test_on",
         "pyspark.pandas.tests.resample.test_error",
+        "pyspark.pandas.tests.resample.test_frame",
         "pyspark.pandas.tests.resample.test_missing",
+        "pyspark.pandas.tests.resample.test_series",
         "pyspark.pandas.tests.resample.test_timezone",
-        "pyspark.pandas.tests.test_frame_resample",
-        "pyspark.pandas.tests.test_series_resample",
         "pyspark.pandas.tests.test_reshape",
         "pyspark.pandas.tests.test_rolling",
         "pyspark.pandas.tests.test_scalars",
@@ -1109,8 +1109,8 @@ pyspark_pandas_connect_part2 = Module(
         "pyspark.pandas.tests.connect.indexes.test_parity_datetime_property",
         "pyspark.pandas.tests.connect.test_parity_frame_interpolate",
         "pyspark.pandas.tests.connect.test_parity_series_interpolate",
-        "pyspark.pandas.tests.connect.test_parity_frame_resample",
-        "pyspark.pandas.tests.connect.test_parity_series_resample",
+        "pyspark.pandas.tests.connect.resample.test_parity_frame",
+        "pyspark.pandas.tests.connect.resample.test_parity_series",
         "pyspark.pandas.tests.connect.test_parity_ewm",
         "pyspark.pandas.tests.connect.test_parity_rolling",
         "pyspark.pandas.tests.connect.test_parity_expanding",
