+++ python/pyspark/pandas/tests/connect/resample/test_parity_frame.py
@@ -16,19 +16,19 @@
 #
 import unittest
 
-from pyspark.pandas.tests.test_frame_resample import FrameResampleTestsMixin
+from pyspark.pandas.tests.resample.test_frame import ResampleFrameMixin
 from pyspark.testing.connectutils import ReusedConnectTestCase
 from pyspark.testing.pandasutils import PandasOnSparkTestUtils, TestUtils
 
 
-class FrameResampleParityTests(
-    FrameResampleTestsMixin, PandasOnSparkTestUtils, TestUtils, ReusedConnectTestCase
+class ResampleParityFrameTests(
+    ResampleFrameMixin, PandasOnSparkTestUtils, TestUtils, ReusedConnectTestCase
 ):
     pass
 
 
 if __name__ == "__main__":
-    from pyspark.pandas.tests.connect.test_parity_frame_resample import *  # noqa: F401
+    from pyspark.pandas.tests.connect.resample.test_parity_frame import *  # noqa: F401
 
     try:
         import xmlrunner  # type: ignore[import]
